Former-commit-id: ed6cbfed234d7a173cc381ef7afbfc944b186a48

+++ src/adapters/Z3ExpressionAdapter.h
@@ -166,11 +166,11 @@ namespace storm {
 					<< ". Integer constant '" << expression->getConstantName() << "' is undefined.";
                 }
                 
-                stack.push(context.int_val(expression->getValue()));    
+                stack.push(context.int_val(static_cast<int>(expression->getValue())));
             }
             
             virtual void visit(ir::expressions::IntegerLiteralExpression* expression) {
-                stack.push(context.int_val(expression->getValueAsInt(nullptr)));
+                stack.push(context.int_val(static_cast<int>(expression->getValueAsInt(nullptr))));
             }
             
             virtual void visit(ir::expressions::UnaryBooleanFunctionExpression* expression) {
